# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API Documentation",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Server Url"
        }
    ],
    "paths": {
        "/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Returns list of users",
                "operationId": "8ab643cdfa2c83702256ca21830b256e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserGet"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store User to database",
                "description": "Store User to database",
                "operationId": "3dde3e31351d54e82d2fc577078f90f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "passwordPanjang"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "passwordPanjang"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserStore"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Get All Users",
                "required": [
                    "status_code",
                    "message",
                    "data"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "string"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    }
                },
                "type": "object"
            },
            "UserGet": {
                "title": "Get All Users",
                "required": [
                    "status_code",
                    "message",
                    "data"
                ],
                "properties": {
                    "status_code": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Data berhasil diambil"
                    },
                    "data": {
                        "$ref": "#/components/schemas/User"
                    },
                    "dev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserStore": {
                "required": [
                    "status_code",
                    "message",
                    "data"
                ],
                "properties": {
                    "status_code": {
                        "type": "integer",
                        "example": 201
                    },
                    "message": {
                        "type": "string",
                        "example": "Data berhasil ditambahkan"
                    },
                    "data": {
                        "$ref": "#/components/schemas/User"
                    },
                    "dev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}